requirements.txt
streamlit
pandas
matplotlib
numpy
plotly.express
json
os
datetime